Document the gitops model
A `gitops/sre` dir is to be used as a starting point for an SRE team to
create their own environment.

An `docs/sre/examples` dir is created to show how the gitops repository
should be used.

The `.tekton` dir has been moved from `pac` to the `template` directory,
and a symlink as been created in the `pac` directory. This is to ensure
that copying the template directory is enough to get an SRE team started.
The link is there to keep the relationship been `.tekton` and `pac`.