Remove argocd sync-wave param from minio
When pipeline-service is imported in infra-deployments project (as
kustomize overlays and not argcd app), the sync-wave params on minio
makes introduces a kind of dead lock. That is tekton-results waiting for
its storage backend and minio waiting for the first sync-wave to finish.
This has been fixed by removing the sync-wave which should still let
minio deploy correctly