Check OLM subscription status
We also need to check the OLM subscription status to verify the install
of Pipeline Service is complete. ArgoCD will sync the OLM Subscription
object, but may not necessarily reflect accurate sync and health data
of the underlying deployments.

Per review, also removing premature exit if an application or
subscription does not meet the install criteria. This allows the
deployment status to be checked/dumped at the end.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>