Add Initial Pipeline Serivce Dashboard
A Grafana dashboard will make it easy for operators to monitor Pipeline
Service components at a glance. The dashboard definitions here can be
deployed using Kustomize on a cluster with the Grafana operator
installed.

This contains an initial dashboard definition, with the intent that
future contributors can add additional graphs and charts. The dashboard
consists of the following:

- Line graphs of completed TaskRuns and PipelineRuns, expressed as a
  rate per hour.
- A success gauge measuring the percent of successful TaskRuns and
  PipelineRuns.

The dashboard assumes that Prometheus will be used as the data source,
and that the following metrics are recorded and exported via a
ServiceMonitor:

- `tekton_pipelines_controller_pipelinerun_count`
- `tekton_pipelines_controller_taskrun_count`