Upgrade PaC version to 0.17.4
As we are working through issues to migrate to the downstream images,
the PaC version has been left unchanged.

There has been quite a few bug fixes that have been delivered that would
improve the UX of RHTAP, therefore an update seems warranted.

The version has been chosen to be compatible with OSP 1.10, so as not to
impact the migration to the downstream image.

Signed-off-by: Romain Arnaud <rarnaud@redhat.com>