Remove usage of chains-ca-cert
The chains-ca-cert ConfigMap was introduced in order to allow the Tekton
Chains controller to push content to the integrated OpenShift registry.
Since then, we have moved away from using this registry altogether in
favor of quay.io.

This change removes usage of the chains-ca-cert ConfigMap as well as the
ArgoCD Jobs responsible for setting it up.

https://github.com/openshift-pipelines/pipeline-service/pull/673

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>