Add Grafana to dev_setup.sh
Add subscription to community Grafana operator and
configure ArgoCD to deploy a Grafana instance using the
mainfests from redhat-appstudio. A separate GrafanaDashboard
instance is created using the in-tree dashboard manifests,
which allows Pipeline Service contributors to iterate on
the dashboard via GitOps. Instructions on how to do "inner
loop" developement are included in a README.