Migrate Tekton-Chains to openshift-pipelines 1/3
This PR leaves the secrets in the tekton-chains namespace, but copies them
to the openshift-pipelines namespace. This should guarantee that older
attestations can still be decoded, and that new attestations can be decoded
if the PR needs to be rolled back.

* Move chains configuration to TektonConfig
* Revert premature changes on namespaces
* Add serviceaccount and job to copy the secrets to the new namespace
* A test commented out has been added for when we are ready to do part 2/2
* Improve debugging of test.sh and behavior on some flakiness

Part 2/3 will enable the public-key-migration test so that EC can target
the new secret.

Part 2/3 will deprecate the tekton-chains namespace.

Signed-off-by: Romain Arnaud <rarnaud@redhat.com>