Fix podman image tag used in CI
We were using the v4.5.1 tag and it's been replaced by v4.6.1. Those
both are floating tags and not static ones, so it does not make much
difference if we start using a bit more generic one v4.6 while still
preserving the versions of the extra rpms installed.
Same happened with the skopeo image, replaced with the new one.