Fix upgrade-tests pipeline
The wrong set of scripts were used when deploying/testing the
baseline branch. Scripts from the baseline branch are now used
when testing the baseline branch, and scripts from the PR are
used when testing the PR branch.

However CI related scripts always use the source branch, so that any
changes in the CI scripts and/or pipeline definitions work as expected.

* Fix issue with a sidecar script that was not called from the sidecar.
* Change target of source code copy to /workspace/source on the sidecar
* Copy sidecar scripts to /workspace/sidecar on the sidecar
* Checkout the revision being tested before setup and test
* Rename ci-runner-container-setup.yaml to setup-ci-runner-container.yaml
so filename and taskname match

Signed-off-by: Romain Arnaud <rarnaud@redhat.com>