update scheduled alert panel to only use successful pipelineruns with scheduled metric
During some recent monitoring of the alert panels in stage and prod, a rash of pipelinerun errors lead to some abnormally high levels for the schedule overhead alert as compared to the execution overhead alert.  This was the motivation to add the status label to the schedule metric like we already have on the other ones.  This change now leverages that label in the grafana dashboard.