Create namespaces before external secrets
ExternalSecrets are creates at sync-wave "-1" and they are required for
services like Tekton Results and OSP. But for the secrets to be created,
the namespace should exist in advance. Here we create the namespace
before the secret. It does not affect the deployment of the OSP
operator, it ignores the fact the namespace it tries to create already
exist.