exporter needs patch permissions
Part of the exporter's new approach to filter throttled pipelineruns because
of the pipeline controller's aggressiver requeue strategy with throttling is that
it labels the pipelinerun for later filtering.  However, the original introduction of
the changes did not add patch permissions to the exporter's RBAC.  This change
addresses that.

rh-pre-commit.version: 2.1.0
rh-pre-commit.check-secrets: ENABLED